pyproject.toml
[project]
name = "health-rater"
version = "0.1.0"
description = "A nutrition score calculator application using the Nutri-Score algorithm"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash>=2.9.0",
    "dash-bootstrap-components>=1.0.0",
    "plotly>=5.0.0",
    "openai>=1.0.0",
    "opencv-python>=4.5.0",
    "pandas>=1.3.0",
    "pillow>=9.0.0",
    "requests>=2.25.0",
    "weasyprint>=55.0", # For PDF export
    "pytz>=2022.1", # For timezone handling
    "pyzbar>=0.1.9",
    "ruff>=0.12.12",
    "langgraph>=0.0.26",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.0.0",
    "black>=22.0.0"
]

[project.scripts]
health-rater = "main:main"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"
